Restructured entire thing, should work now (-fonts)
