--- archaeology/tours/views.py
@@ -1,5 +1,0 @@
-from django.shortcuts import render
-
-# Create your views here.
-def index(request):
-    return render(request, 'tours/index.html')
