+++ archaeology_tours/berry/views.py
@@ -1,0 +1,8 @@
+from django.shortcuts import render
+from django.http import HttpResponse
+from django.template import loader
+
+# Create your views here.
+def index(request):
+   template = loader.get_template('index.html')
+   return HttpResponse(template.render())
