--- first.txt
@@ -1,2 +1,0 @@
-github is awesome
-hi
